Add helpers for reading .NET crazy-endian UUIDs
This adds a helper class containing methods capable of producing a Java
UUID type from a byte array or Netty ByteBuf containing a UUID
formatted with the .NET Guid.ToByteArray() method which produces
something affectionately (!) called crazy-endian. It also adds tests for
these functions.

We pull in JUnit 4 as a testing framework in the test scope.